xuantie: pmu: Support cycles&instructions overflow interrupt
The Sscofpmf spec does not define overflow interrupts for cycles & instructions [1].
But this is indispensable in performance profiling.

Xuantie added registers mhpmevent0 and mhpmevent2, with addresses 0x7E0 and 0x7E1
respectively, which are used to configure cycles & instructions.

At the same time, mxstatus[8] is newly defined as ofint to control whether the
cycles and instructions events of HPCP can generate interrupt 13 (lcofip). When
it is 1, an interrupt can be generated, and when it is 0, an interrupt cannot be
generated.

This patch is used to support the above functions.

[1] https://github.com/riscvarchive/riscv-count-overflow/issues/5

Signed-off-by: Chen Pei <cp0613@linux.alibaba.com>